src/requirements.txt
aiohttp==3.9.3
azure-storage-blob==12.20.0
botbuilder-core==4.15.0
botbuilder-integration-aiohttp==4.15.0
cloudpickle==3.0.0
faiss-cpu==1.8.0
langchain-core==0.2.12
langchain-openai==0.1.14
langchain-community==0.2.7
langchain-text-splitters==0.2.2
msal==1.28.0
numpy==1.26.4
openai==1.35.10
pandas==2.2.2
python-dotenv==1.0.1
tiktoken==0.7.0
tqdm==4.66.2
